==> 3_Topic_modeling/1.Preprocessing_TFIDF.ipynb <==
import pandas as pd 
import numpy as np 
import joblib
import string
from stemming.porter2 import stem
from nltk.corpus import stopwords
import operator

### Download the database with one row for each publication

df = pd.read_csv("../2_Treatment_database/output/database_one_row_each_paper.csv")
print( "Loaded %d X %d dataframe with unique papers" % (len(df), len(df.columns) ))

### Preprocess the abstracts

Transform abstracts in lower cases

abstracts = df['abstract'].str.lower().to_list()

Remove punctuation and replace it by on white space and then replace double white by one white space

string.punctuation =  string.punctuation + '’—☆–−()()©“”‘'
table_remove_punct_white_space = str.maketrans(string.punctuation, ' '*len(string.punctuation))
abstracts = [abstract.translate(table_remove_punct_white_space) for abstract in abstracts]
abstracts = [' '.join(abstract.split()) for abstract in abstracts]

Define stopwords 
* nltk stopwords
* countries and demonyms 

nltk_stopwords = list(set(stopwords.words('english')))

countriesPath = "../0_Reference_files/UNSD_database.csv"
countries_df = pd.read_csv(countriesPath)
countries_list = countries_df['Country0'].tolist()
demonym1_list = countries_df['Demonym1'][countries_df['Demonym1'].notnull()].tolist()
demonym2_list = countries_df['Demonym2'][countries_df['Demonym2'].notnull()].tolist()
names_list = countries_list + demonym1_list + demonym2_list

names_list = [str(name).lower().translate(table_remove_punct_white_space) for name in names_list]
names_list = [' '.join(name.split()) for name in names_list]

Remove names that are not single word

not_single = [name for name in names_list
                     if (len(name)>3 & len(name.split())>1)]

l = [''] *len(not_single)

trans = {name:delet for name,delet in zip(not_single,l)}

for a, b in trans.items():
    for i in range(len(abstracts)):
        abstracts[i] = abstracts[i].replace(a, b)

custom_stopwords = nltk_stopwords + names_list

Remove stopwords and group words by stem :

abstracts = [" ".join([stem(word) for word in abstract.split(" ") 
                       if ((word not in custom_stopwords) & (len(word)>2))]) for abstract in abstracts]

Remove stem stopwords 
* synonyms of mitigation
* that appeared useless after topic modeling experiences:

stem_stopwords = ['mitig','carbon','decarbonis','decarbon','co2','ghg','greenhous','gas','emiss','reduc','reduct']
stem_stopwords += ['million','ton','billion','proceed','paper','data','present','result','studi','describ','also','base','explor','analys','analyz','elsevi']

abstracts = [" ".join([stem for stem in abstract.split(" ") if stem not in stem_stopwords]) for abstract in abstracts]

Remove digits

abstracts =[" ".join([stem for stem in abstract.split(" ") if not stem.isdigit()]) for abstract in abstracts]

Save processed abstracts

abstract_prepro = pd.DataFrame(abstracts, columns = ['abstracts_prepro'])

abstract_prepro.to_csv("./interm/processed_abstracts.csv",index=False)

### Term weighting with TF-IDF

from sklearn.feature_extraction.text import TfidfVectorizer

tfidf_vectorizer = TfidfVectorizer(max_df=0.95, min_df=0.01,smooth_idf= False)
                                   
tfidf = tfidf_vectorizer.fit_transform(abstracts)
print('Created %d X %d TF-IDF-normalized document-term matrix' % (tfidf.shape[0], tfidf.shape[1]))

# extract the resulting vocabulary
tfidf_feature_names = tfidf_vectorizer.get_feature_names()
print("Vocabulary has %d distinct terms" % len(tfidf_feature_names))

Save TF-IDF matrix and vocabulary features names under pkl format

joblib.dump((tfidf, tfidf_feature_names), "./interm/tfidf_matrix-features_names.pkl") 

### Look at weightest words

def rank_terms(tfidf, terms):
    # get the sums over each column
    sums = tfidf.sum(axis=0)
    # map weights to the terms
    weights = {}
    for col, term in enumerate(terms):
        weights[term] = sums[0,col]
    # rank the terms by their weight over all documents
    return sorted(weights.items(), key = operator.itemgetter(1), reverse=True)

ranking = rank_terms(tfidf,tfidf_feature_names)
for i, pair in enumerate(ranking[0:20]):
    print( "%02d. %s (%.2f)" % ( i+1, pair[0], pair[1] ) )
