==> 3_Topic_modeling/5. Association_topics_abstracts.ipynb <==
import pandas as pd
import numpy as np
import joblib

### Download data and models
#### Database with additional treatment

df = pd.read_csv("../2_Treatment_database/output/database_one_row_each_paper.csv")
print( "Loaded %d X %d dataframe with unique papers" % (len(df), len(df.columns) ))

#### Preprocessed abstracts

abstracts_prepro = pd.read_csv("./interm/processed_abstracts.csv")
print( "%d preprocessed abstracts" % (len(abstracts_prepro)) )

#### TF-IDF normalised document-term matrix and terms

(tfidf, tfidf_feature_names) = joblib.load( "./interm/tfidf_matrix-features_names.pkl" )
print( "Loaded %d X %d document-term matrix" % (tfidf.shape[0], tfidf.shape[1]) )

#### Model with optimal parameters:

k_topics,alpha,l1,W,H = joblib.load("./output/model_selected.pkl") 
param = [k_topics,alpha,l1,W,H]
param_names = ["k_topics", "alpha", "l1"]
for name in param_names:
    print(name, '=', repr(eval(name)))

### Run the model

WH = W.dot(H)
WH.shape

W.shape

### Normalize W 
(the sum of coefficients of each row is equal to 1)

W_norm = np.zeros(W.shape)
for k in range(k_topics):
    topic_index = k
    sum_terms_topic = np.sum(H[topic_index,:])
    for n in range(len(df)):
        abstract_index = n
        sum_terms_abstract = np.sum(WH[abstract_index,:])
        abstract_topic = W[abstract_index,topic_index] * sum_terms_topic
        W_norm[abstract_index,topic_index] = abstract_topic/sum_terms_abstract

### Create topics columns

def get_5_words( all_terms, H, topic_index):
    # reverse sort the values to sort the indices
    top_indices = np.argsort( H[topic_index,:] )[::-1]
    top_terms=[]
    for i in range(5):
        term = tfidf_feature_names[top_indices[i]]
        top_terms.append(term)
    return top_terms

columns_topics = []
list_words_topics = []
for k in range(k_topics):
    words_topics = get_5_words(tfidf_feature_names, H , k)
    list_words_topics.append(words_topics)
    column_name = str(k) + '_'+ "_".join(words_topics)        
    columns_topics.append(column_name)

### Create a dataframe (df_binaire) with binary topic variables (1 if normalized coefficient $\geq$ 0.02, else 0)

b = (W_norm>0.02).astype(int)
df_binaire = pd.DataFrame(data = b)
#Create row Total: 
df_binaire.loc['Total',:]= df_binaire.sum(axis=0)
#Create column Total: 
df_binaire.loc[:,'Total'] = df_binaire.sum(axis=1)

### Create a dataframe (prop) containing the number of publications per topic

prop = pd.DataFrame()
prop['topic']= columns_topics
prop['seuil_0.02']=np.nan
prop.set_index('topic', inplace = True)
for i in range(k_topics):
    row = columns_topics[i]
    prop.loc[row,['seuil_0.02']] = df_binaire.loc['Total',[i]].values[0]

#### Rename columns with the five weightest terms charactering each topic

columns_topics.append('Total_topics_in_paper')

df_binaire.columns = columns_topics
df_binaire.head(1)

### How many publications per topic

prop

### How many topics characterize how many papers

Remove "Total" row

df_binaire = df_binaire.drop(['Total'])

df_paper = pd.DataFrame(df_binaire.Total_topics_in_paper.value_counts(), columns=['Total_topics_in_paper'])
df_paper

### Add processed abstracts to the main database

df['stem_abstract'] = abstracts_prepro["abstracts_prepro"]

### Add topics attribution to the main database

df = df.reindex(columns= np.append(df.columns,columns_topics))
df.update(df_binaire)

df.head(2)

df.to_csv('./output/database_seuil_0.02.csv',index=False)
